third_party/WebKit/LayoutTests/platform/win/media/autoplay-muted-datasaver-on-expected.txt

CONSOLE WARNING: line 23: Failed to execute 'play' on 'HTMLMediaElement': API can only be initiated by a user gesture.
This is a testharness.js-based test.
PASS Test muted autoplay videos interaction with data saver off 
Harness: the test ran to completion.
